add `.worktrees/` to `.gitignore`
the `.worktrees/` folder is used in the guide for manual backporting as
the location where to create the worktree for the backport branch.

as this is inside of the repository the folder will be reported as an
untracked file. to avoid this, i've now added it to `.gitignore`.

though it might be better to create the worktree outside of the
repository, the documentation on it doesn't comment explicitly on using
it inside of the main worktree, however none of the examples do that.

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>