copy `build.sh` over from opensearch-build
the build script should be located in this repository instead of
centrally in [opensearch-build][] to allow per-branch specific builds
(newer versions might have newer requirements, which currently cannot be
reflected).

the file was copied 1:1 from `scripts/components/OpenSearch/build.sh` as
of commit 9e1aa38 in the [opensearch-build][] repository.

this is part of opensearch-build#99 and needs to be backported to all
active branches because the file here will only be used by the build
once the central one in opensearch-build has been removed which in turn
can only happen when all active branches have the `build.sh`.

[opensearch-build]: https://github.com/opensearch-project/opensearch-build

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>