build no-jdk distributions as part of release build
with commit e3572464859 the `build.sh` has been moved to this repository
and with commit dfa1118eee8 the missing no-jdk builds have been added.
with these preliminary works done we can now start building the no-jdk
distributions as part of the release build so that they can then be
published.

this is part of opensearch-project/opensearch-build#99.

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>