[Manual] Revert Add bwcVersion 2.5.0
In the middle of getting ready for the 2.4.0 release our backward
compatibility tests for security started failing, which looks to be
related to this change.

Reverting this commit to unblock pull request checks in to main.

How we can stage changes like this before the 2.5.0 build
is ready, or do we need to careful make these changes when main branches
won't be impacted?

Closes https://github.com/opensearch-project/OpenSearch/issues/5076

Signed-off-by: Peter Nied <petern@amazon.com>