Fix test failing due to too long http line exception
The `sendBodyAsSourceParam` method determines whether it should send the body as query parameter. It used to check the content length of the entity which in some cases (e.g. regex expression) is much shorter than the encoded-url and may result in a TooLongHttpLineException.

This commit makes an additional check on the length of the url-encoded string. To avoid unnecessary computation it is only checked when the RandomizedTest.rarely() returns true.

Signed-off-by: Rabi Panda <adnapibar@gmail.com>