Fix 1.x compatibility bug with stored Tasks (#5412)
When the new 'cancelled' field was introduced it was a miss not to
increment the version number on the mapping definitions for the .tasks
index. This commit fixes that oversight, as well as modifies the
existing backward compatiblity test to ensure that it will catch future
mistakes like this one.

Closes #5376

Signed-off-by: Andrew Ross <andrross@amazon.com>
(cherry picked from commit 4616dfac5382062a5149c799554dda4ff601369f)
Signed-off-by: Andrew Ross <andrross@amazon.com>