Fix flaky SR test testStartReplicaAfterPrimaryIndexesDocs.
This test was failing because we are validating post recovery if a shard is able to perform segrep while also performing validation if a passed in checkopint.  In the post recovery test this checkpoint is always empty, yet the shard will be ahead of this checkpoint after docs are indexed.  This change differentiates shard validation from checkpoint validation.

Signed-off-by: Marc Handalian <handalm@amazon.com>

Fix spotless.

Signed-off-by: Marc Handalian <handalm@amazon.com>

Fix testIsSegmentReplicationAllowed_WrongEngineType.

Signed-off-by: Marc Handalian <handalm@amazon.com>

Update warn logs in isSegmentReplicationAllowed.

Signed-off-by: Marc Handalian <handalm@amazon.com>