Fix flaky test IndexStatsIT.testFilterCacheStats.
This change disables query cache when internally querying for operation uid.
This code path is executed only when assertions are enabled to check if an op already exists in the index.
The current code fetches the searcher using acquireSearcher, which reuses the queryCache and conflicts with tests on the cache.

Signed-off-by: Marc Handalian <handalm@amazon.com>