[Refactor] core.common to new opensearch-commons library
Refactors all of o.opensearch.common classes in the core library to a new
opensearch-commons library (inspired by apache commons). The intent is to
refactor any o.opensearch.common.* classes in :server module to this new
library. This will be done with care such that the dependencies are carefully
managed to avoid any cyclic or unnecessarily complicated dependencies.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>