Fix flaky SegmentReplicationITs.
This change fixes flakiness with segment replication ITs.
It does this by updating the wait condition used to ensure replicas are up to date
to wait until a searched docCount is reached instead of output of the Segments API that can change
if there are concurrent refreshes.
It also does this by updating the method used to assert segment stats to wait until the assertion holds true rather
than at a point in time.  This method is also updated to assert store metadata directly over API output.

Signed-off-by: Marc Handalian <handalm@amazon.com>