Update DEVELOPER_GUIDE.md to document bwc and feature flag mechanims
OpenSearch uses @opensearch.internal, @opensearch.api, @opensearch.experimental,
and feature flags to signal maturity and usage intent of the various
implementation components. Each one of these annotations carry a level of
backwards compatibility guarantees and support requirements that is not
officially documented. This change officially documents those bwc requirements,
level of guarantee, and recommended approaches in how to best use the mechanisms
to communicate this to the development and user community.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>