Fix potential deadlock in ConcurrentInvocationLinearizer
There is potential for a deadlock if work is submitted using the same
executor as provided to ConcurrentInvocationLinearizer. This commit
changes the behavior of that class to not implement any concurrency
itself, therefore avoiding any potential for deadlock. The purpose of
this class is to share the result when concurrent threads ask for the
same resource, not provide concurrency itself, so I think the change is
appropriate.

I've also removed the future from the TransferManager API as the only
caller of it immediately blocked on the result. We may need to implement
asynchronous logic here in the future, but I think it will be better to
design that when there is a use case and a synchronous method makes
things simpler now. I've also added logic to ensure exceptions get
properly unwrapped from ExecutionException.

Signed-off-by: Andrew Ross <andrross@amazon.com>