Use per-key latch to wait on file downloads
The lock that guards `FileCache.compute` is per-cache-segment, and
therefore means unrelated keys can get stuck waiting for one another.
This refactors the code to do the download outside of the cache
operation, and uses a per-key latch mechanism to ensure that only
requests for the exact same blob will block on each other.

See [this issue][1] for details about the cache implementation. I think
it is possible to re-work the cache so that locking would be much more
precise and this change would not be necessary. However, that is a
bigger change potentially with other tradeoffs, so I think this fix is a
reasonable thing to do now.

[1]: https://github.com/opensearch-project/OpenSearch/issues/6225#issuecomment-1496337366

Signed-off-by: Andrew Ross <andrross@amazon.com>