[Refactor] SegmentsStats#filesSizes from ImmutableOpenMap to java.util.Map
Using an ImmutableOpenMap in filesSizes in o.o.index.engine.SegmentsStats
is overkill since all we care about is public facing API not leaking the
filesSizes collection. This commit refactors the SegmentsStats#filesSizes
variable from an ImmutableOpenMap to an unmodifiable java.util.Map to
remove all of the unnecessary temporary object copies on the internal
map.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>