Make file cache stats consistent with other APIs
Some cosmetic changes here to make the stat names consistent with
[RequestCacheStats][1] and [QueryCacheStats][2] that already exist in
the API.

I've also removed the 'replaced' and 'removed' stats. These are internal
cache details that I don't think will be useful to a user. We can always
add them back later if needed.

Note this is backward incompatible, but this is okay as this will get in
ahead of the next release which removes the feature flag from searchable
snapshots.

[1]: https://github.com/opensearch-project/OpenSearch/blob/main/server/src/main/java/org/opensearch/index/cache/request/RequestCacheStats.java
[2]: https://github.com/opensearch-project/OpenSearch/blob/main/server/src/main/java/org/opensearch/index/cache/query/QueryCacheStats.java

Signed-off-by: Andrew Ross <andrross@amazon.com>