Fix snapshotting searchable snapshot indexes (#7247)
Two changes here: First, when snapshotting a searchable snapshot index
(or snapshotting a full cluster that contains searchable snapshot
indices) then we will snapshot the index metadata which includes the
pointer to the original snapshot, but skip copying the data since it
already exists in the source snapshot. Second, when restoring an index
from a snapshot that is itself a searchable snapshot index, then it must
be handled as such and restored as a searchable snapshot index.

Resolves #7204

Signed-off-by: Andrew Ross <andrross@amazon.com>
(cherry picked from commit e1da84d9d2e648066f3948906f935ce06495aac4)