[Refactor] Version and LegacyESVersion from server module to core lib (#7328)
Refactors the Version and LegacyESVersion classes from the server module
to the core library. All gradle build scripts and dependencies are
updated. This commit serves as a foundation for refactoring all
OpenSearch base classes from the server module into appropriate
libraries to decouple core search logic from the cluster model so
it can be reused in serverless and cloud native environments.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>