Rewrite version update GHA to commit and push
This change alters the version workflow to create a new issue, then
create several commits linked back to that issue and finally close the
issue when all commits are completed for tracability of version updates.

Background

As version updates have different propigation speeds depending on pull
request reviews, consistant CI in many branches, and an careful order of
execution this would cause failures outside the scope of changes that
would block changes into main and other branches.

This change removes the use of pull requests and the standard ci gates.
To ensure that the changes operate consistantly created a GitHub Action
to manage the file modification peternied/opensearch-core-version-updater

See an example of this workflows results with https://github.com/peternied/OpenSearch-1/issues/66

Related issues
- Resolves #7411
- Related Discussion #7398
- Resolves #7396

Signed-off-by: Peter Nied <petern@amazon.com>