Segment Replication - Force segrep round on replicas once recovery completes
This change fixes a rare case a replica can be left stale if its primary does not publish a new checkpoint
after the replica completes recovery.  The replica will perform a forced round of segment replication as part of
the recovery process to catch up, however if it receives a new checkpoint during that time it will be dropped. If subsequently the primary stops indexing/writing new segments, the replica is left stale.

Signed-off-by: Marc Handalian <handalm@amazon.com>