Segment Replication - Fixes for test SegmentReplicationIT.testReplicaHasDiffFilesThanPrimary (#7513)
* Segment Replication - Force segrep round on replicas once recovery completes

This change fixes a rare case a replica can be left stale if its primary does not publish a new checkpoint
after the replica completes recovery.  The replica will perform a forced round of segment replication as part of
the recovery process to catch up, however if it receives a new checkpoint during that time it will be dropped. If subsequently the primary stops indexing/writing new segments, the replica is left stale.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Fix all shards failed.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Ensure NRT engine is open before triggering refresh.

Without this check tests will fail because RefreshMetricUpdater#afterRefresh will
get invoked without a call to beforeRefresh. This happens when readerManager.maybeRefresh is
invoked on a closed engine.

Signed-off-by: Marc Handalian <handalm@amazon.com>

Fix NPE caused when getLatestReplicationCheckpoint is returned as null.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Ensure segrep is enabled on index for SegmentReplicationTargetService's IndexEventListener implementations.

Signed-off-by: Marc Handalian <handalm@amazon.com>

Add integration test.

Signed-off-by: Marc Handalian <handalm@amazon.com>

Spotless.

Signed-off-by: Marc Handalian <handalm@amazon.com>

Fix testDropPrimaryDuringReplication

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Remove unused NO_OP versions of service components.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Add more unit tests.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* PR feedback.

Signed-off-by: Marc Handalian <handalm@amazon.com>

---------

Signed-off-by: Marc Handalian <handalm@amazon.com>