Generic Streamable Registry
This commit adds a WriteableRegistry to register generic read / write
serialization methods for StreamInput.readGeneric and
StreamOutput.writeGeneric. This enables modules, plugins, and libraries
to register their own type serialization logic without forcing the class
implementation to be in the :server module.

Decoupling this logic also further decouples the StreamInput /
StreamOutput transport logic from the server module such that it can be
refactored to the core library to support cloud native and serverless
implementations.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>