[Refactor] StreamInput, StreamOutput, Writeable to core library
This is a big commit as it refactors the StreamInput and StreamOutput
core serialization and supporting classes from the server module to the
core library such that it can be reused and overridden in cloud native
and serverless applications. The majority of the changes in this commit
are the import statements. Noteable refactors include:

* StreamInput
* StreamOutput
* Writeable
* NamedWriteable
* NamedWriteableRegistry
* Index
* ShardId
* BytesReference
* ByteArray
* BigArray
* SecureString
* Text
* ParsingException
* RestStatus

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>