Named queries should return their internal score as well
While some internal queries may not be scored or not contrbute to the
final doc score, in some cases (such as knn queries compbined with other
queries) the distance metric is very interesting and should be reported
back as part of the response.

This change adds that by default, so named queries always return the
score (instead of just the name). It adds no additional cost.

Signed-off-by: Itamar Syn-Hershko <itamar@bigdataboutique.com>