Remove error path for ancient compression detection
The error message generated by this code path refers to Elasticsearch
version 2.0.0.beta1, which is likely to cause confusion with the
OpenSearch 2.x line if it were ever to be encountered. The ES2 version
is so old that I don't think we need any special handling here and can
just fall back to the default "not xcontent" error message.

Signed-off-by: Andrew Ross <andrross@amazon.com>