Fix testReplicaHasDiffFilesThanPrimary.
This test is failing in two ways.
First it fails when copying segments from the remote store and there is a cksum mismatch. In this case
it is not guaranteed the directory implementation will replace the existing file when copying from the store.  This change ensures the mismatched file is cleaned up but only if the shard is not serving reads.  In that case we fail the shard so it is re-recovered rather than deleting the segment underneath it.

This test also fails with a divide by 0 in RemoteStoreRefreshListener.

Signed-off-by: Marc Handalian <handalm@amazon.com>