[Bug] Fix Setting#byteSizeSetting to use whole value string (#9011)
This fixes a bug where the Setting#byteSizeSetting static method
returned fractional byte values when calling .toString inside the
lambda. The change is to directly return the bytes value as a string
instead of using ByteSizeValue#toString which tries to be smart and
by pretty printing in a more human-friendly format.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>
Signed-off-by: Ivan Brusic <ivan.brusic@flocksafety.com>